ml/pyproject.toml
[project]
name = "riya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "pymongo>=4.11",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.5",
    "scikit-learn>=1.6.1",
    "textblob>=0.19.0",
]
